speech-to-text/requirements.txt
aniso8601==8.1.0
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-RESTful==0.3.8
future==0.18.2
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.4
google-cloud-speech==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.15
MarkupSafe==1.1.1
mypy-extensions==0.4.3
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
